+++ conversejs/lib/@types/global.d.ts
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2024 John Livingston <https://www.john-livingston.fr/>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+// Important note: loc segments that are declared here must also be in loc.keys.js (for now).
+
+declare const LOC_ANNOUNCEMENTS_MESSAGE_TYPE_STANDARD: string
+declare const LOC_ANNOUNCEMENTS_MESSAGE_TYPE_ANNOUNCEMENT: string
+declare const LOC_ANNOUNCEMENTS_MESSAGE_TYPE_HIGHLIGHT: string